Core/Pets: Adjust pet slot saving logic after constant changes
Port From (https://github.com/TrinityCore/TrinityCore/commit/fd9eb28322fd75e94fcf5093aaf54bee145edfe8)

+++ Source/Game/Entities/Player/Player.DB.cs
@@ -1558,7 +1558,7 @@ namespace Game.Entities
                     petInfo.Level = result.Read<byte>(3);
                     petInfo.Experience = result.Read<uint>(4);
                     petInfo.ReactState = (ReactStates)result.Read<byte>(5);
-                    PetSaveMode slot = (PetSaveMode)result.Read<byte>(6);
+                    PetSaveMode slot = (PetSaveMode)result.Read<short>(6);
                     petInfo.Name = result.Read<string>(7);
                     petInfo.WasRenamed = result.Read<bool>(8);
                     petInfo.Health = result.Read<uint>(9);
